feat: 2 pow 32 bytes guest memory address (#3104)
Adds support for 2^32 bytes memory address. Supersedes #2850.

Benchmark:
https://github.com/axiom-crypto/openvm-eth/actions/runs/31222063368

New design:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32282531103/job/96164388705

Bench after MEM_BITS change:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32427802054/job/96613241136

After fixing some codex audit findings:
https://github.com/axiom-crypto/openvm-eth/actions/runs/32431879330

Closes INT-8080

---------

+++ crates/continuations/src/tests/e2e.rs
@@ -327,7 +327,7 @@ fn test_deferral_e2e() -> Result<()> {
 
     let elf = Elf::decode(
         include_bytes!("../../programs/examples/multiple.elf"),
-        MEM_SIZE as u32,
+        MEM_SIZE as u64,
     )?;
     let exe = VmExe::from_elf(
         elf,
